Finshed the initial addition of all the new multicast code with significant help from Satya and Venu.
merged OpenSLP 1.0.5

merged OpenSLP 1.0.4

+++ openslp/libslp/libslp_parse.c
@@ -438,7 +438,7 @@ SLPError SLPParseAttrs(const char* pcAttrList,
 {
     const char* slider1;
     const char* slider2;
-    size_t      attridlen;
+    int			attridlen;
 
     /* Check for bad parameters */
     if(pcAttrList == 0 ||
